requirements.txt
google-cloud-firestore==2.5.3
beautifulsoup4==4.11.1
